# static/openapi.json
{
  "info": {
    "title": "Book Inventory REST API",
    "description": "A web api to handle book inventory operations.",
    "version": "v1.1"
  },
  "openapi": "3.0.3",
  "servers": [
    {
      "url": "https://docker-book-inventory-api.onrender.com"
    },
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "tags": [
    {
      "name": "books",
      "description": "book operation"
    },
    {
      "name": "students",
      "description": "student"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Returns All books in database",
        "description": "Returns All books in database",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "book": {
                      "type": "object",
                      "example": [
                        {
                          "author": "Leo Tolstoy",
                          "genre": "Comedy",
                          "isbn": "512-135-134-879-1",
                          "title": "ZECTAA",
                          "year_published": 2000
                        },
                        {
                          "author": "Mark",
                          "genre": "Comedy",
                          "isbn": "123-14132-5123",
                          "title": "Funnnnnn Book",
                          "year_published": 2022
                        }
                      ]
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Add new book",
        "description": "Add new book",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/book"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "book <Book Title> has been added successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "An Error Occured While Inserting The Book"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Return specific book",
        "description": "Return specific book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id number",
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "serial4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "book": {
                      "type": "object",
                      "example": {
                        "author": "Tastoy",
                        "genre": "Comedy",
                        "isbn": "222-135-134-879-1",
                        "title": "Caaer 2",
                        "year_published": 2000
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "books"
        ],
        "summary": "Edit specific book",
        "description": "Edit specific book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id number",
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "serial4"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Book <Book Title> successfully updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "books"
        ],
        "summary": "Delete a book by id",
        "description": "Delete a book by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id number",
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "serial4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Book <Book Title> successfully deleted."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      }
    },
    "/books/borrow": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Returns All books borrowed in database",
        "description": "Returns All books borrowed in database",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "students": {
                      "type": "object",
                      "example": [
                        {
                          "Student Name": "Zack Queen",
                          "author": "Jane Austen",
                          "genre": "Epic",
                          "isbn": "978-366-98371-770-0",
                          "title": "To Kill a Mockingbird",
                          "year_published": 1927
                        },
                        {
                          "Student Name": "Malik",
                          "author": "Harper Lee",
                          "genre": "Romance",
                          "isbn": "978-156-31577-242-6",
                          "title": "War and Peace",
                          "year_published": 1911
                        }
                      ]
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      }
    },
    "/books/borrow/{id}": {
      "put": {
        "tags": [
          "books"
        ],
        "summary": "Update Student ID on a book",
        "description": "Update Student ID on a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id number",
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "serial4"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "student_id": {
                    "type": "string",
                    "example": "99257aa5-5e60-4217-be55-d87cd167445f"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book {book.title} borrow status has been updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      }
    },
    "/students": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Returns All books in database",
        "description": "Returns All books in database",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "students": {
                      "type": "object",
                      "example": [
                        {
                          "student_id": "7e8c29cf-95c8-4450-816d-69d16344f1c9",
                          "name": "Malik",
                          "email": "malikk11@hotmail.com",
                          "membership_start": "Sat, 25 Dec 2021 00:00:00 GMT"
                        },
                        {
                          "student_id": "7e8c29cf-95c8-4450-816d-69d16344f1c9",
                          "name": "Malik",
                          "email": "malikk11@hotmail.com",
                          "membership_start": "Sat, 25 Dec 2021 00:00:00 GMT"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "students"
        ],
        "summary": "Add New Student",
        "description": "Add New Student",
        "requestBody": {
          "description": "description->None",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/student"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Student Zack King has been added successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "book": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "format": "varchar",
            "example": "978-554-72587-216-7"
          },
          "title": {
            "type": "string",
            "format": "varchar",
            "example": "Book of bok"
          },
          "author": {
            "type": "string",
            "format": "varchar",
            "example": "Da vin ci"
          },
          "year_published": {
            "type": "integer",
            "format": "int4",
            "example": 1990
          },
          "genre": {
            "type": "string",
            "format": "varchar",
            "example": "Comedy"
          }
        }
      },
      "student": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Zack King"
          },
          "membership_start": {
            "type": "string",
            "example": "2012-05-20T00:00:00.000Z"
          },
          "email": {
            "type": "string",
            "example": "freemail@gmail.com"
          }
        }
      }
    }
  }
}